DO* version of LIST-FREE-VARIABLES.
Comparing to LOOP version, this version is better because we can write
the epilogue process in a very straightforward manner and that
contribute to avoid the additinal cost to APPEND the zero vector at the
*end* of the given matrix. Since matrices is expressed as a list of
row-vectors, the cost of that process gains linearly to the size of
matrix.

The initialization and update of the variable ROW and NEXT are postponed
until the BODY of DO*, or we need to write identical expression twice.
Even though the explicit use of SETF is not preferable, the same
expression written twice is harmful and should be avoided.